// src/types/auto-generated.d.ts
export interface components {
  schemas: {
// ...
    HostnameGeneratorItem: {
// ...
      spec: {
        selector?: {
          meshExternalService?: {
            matchLabels?: {
              [key: string]: string;
            };
          };
          meshService?: {
            matchLabels?: {
              [key: string]: string;
            };
          };
        };
        template?: string;
      };
    };
    HostnameGeneratorCreateOrUpdateSuccessResponse: {
      /**
       * @description warnings is a list of warning messages to return to the requesting Kuma API clients.
       * Warning messages describe a problem the client making the API request should correct or be aware of.
       */
      warnings?: string[];
    };
    MeshExternalServiceItem: {
// ...
      labels?: {
        [key: string]: string;
      };
      /** @description Spec is the specification of the Kuma MeshExternalService resource. */
      spec: {
// ...
        tls?: {
          /**
           * @description AllowRenegotiation defines if TLS sessions will allow renegotiation.
           * Setting this to true is not recommended for security reasons.
           * @default false
           */
          allowRenegotiation?: boolean;
          /**
           * @description Enabled defines if proxy should originate TLS.
           * @default false
           */
          enabled?: boolean;
          /** @description Verification section for providing TLS verification details. */
          verification?: {
            /** @description CaCert defines a certificate of CA. */
            caCert?: {
              /**
               * Format: byte
               * @description Data source is inline bytes.
               */
              inline?: string;
              /** @description Data source is inline string` */
              inlineString?: string;
              /** @description Data source is a secret with given Secret key. */
              secret?: string;
            };
            /** @description ClientCert defines a certificate of a client. */
            clientCert?: {
              /**
               * Format: byte
               * @description Data source is inline bytes.
               */
              inline?: string;
              /** @description Data source is inline string` */
              inlineString?: string;
              /** @description Data source is a secret with given Secret key. */
              secret?: string;
            };
            /** @description ClientKey defines a client private key. */
            clientKey?: {
              /**
               * Format: byte
               * @description Data source is inline bytes.
               */
              inline?: string;
              /** @description Data source is inline string` */
              inlineString?: string;
              /** @description Data source is a secret with given Secret key. */
              secret?: string;
            };
            /**
             * @description Mode defines if proxy should skip verification, one of `SkipSAN`, `SkipCA`, `Secured`, `SkipAll`. Default `Secured`.
             * @default Secured
             * @enum {string}
             */
            mode?: "SkipSAN" | "SkipCA" | "Secured" | "SkipAll";
            /** @description ServerName overrides the default Server Name Indicator set by Kuma. */
            serverName?: string;
            /** @description SubjectAltNames list of names to verify in the certificate. */
            subjectAltNames?: ({
                /**
                 * @description Type specifies matching type, one of `Exact`, `Prefix`. Default: `Exact`
                 * @default Exact
                 * @enum {string}
                 */
                type?: "Exact" | "Prefix";
                /** @description Value to match. */
                value: string;
              })[];
          };
          /** @description Version section for providing version specification. */
          version?: {
            /**
             * @description Max defines maximum supported version. One of `TLSAuto`, `TLS10`, `TLS11`, `TLS12`, `TLS13`.
             * @default TLSAuto
             * @enum {string}
             */
            max?: "TLSAuto" | "TLS10" | "TLS11" | "TLS12" | "TLS13";
            /**
             * @description Min defines minimum supported version. One of `TLSAuto`, `TLS10`, `TLS11`, `TLS12`, `TLS13`.
             * @default TLSAuto
             * @enum {string}
             */
            min?: "TLSAuto" | "TLS10" | "TLS11" | "TLS12" | "TLS13";
          };
        };
      };
      /** @description Status is the current status of the Kuma MeshExternalService resource. */
      status?: {
        /** @description Addresses section for generated domains */
        addresses?: {
            hostname?: string;
            hostnameGeneratorRef?: {
              name: string;
            };
            origin?: string;
          }[];
        hostnameGenerators?: ({
            /** @description Conditions is an array of hostname generator conditions. */
            conditions?: ({
                /**
                 * @description message is a human readable message indicating details about the transition.
                 * This may be an empty string.
                 */
                message: string;
                /**
                 * @description reason contains a programmatic identifier indicating the reason for the condition's last transition.
                 * Producers of specific condition types may define expected values and meanings for this field,
                 * and whether the values are considered a guaranteed API.
                 * The value should be a CamelCase string.
                 * This field may not be empty.
                 */
                reason: string;
                /**
                 * @description status of the condition, one of True, False, Unknown.
                 * @enum {string}
                 */
                status: "True" | "False" | "Unknown";
                /**
                 * @description type of condition in CamelCase or in foo.example.com/CamelCase.
                 * ---
                 * Many .condition.type values are consistent across resources like Available, but because arbitrary conditions can be
                 * useful (see .node.status.conditions), the ability to deconflict is important.
                 * The regex it matches is (dns1123SubdomainFmt/)?(qualifiedNameFmt)
                 */
                type: string;
              })[];
            hostnameGeneratorRef: {
              name: string;
            };
          })[];
        /** @description Vip section for allocated IP */
        vip?: {
          /** @description Value allocated IP for a provided domain with `HostnameGenerator` type in a match section. */
          ip?: string;
        };
      };
    };
    MeshExternalServiceCreateOrUpdateSuccessResponse: {
      /**
       * @description warnings is a list of warning messages to return to the requesting Kuma API clients.
       * Warning messages describe a problem the client making the API request should correct or be aware of.
       */
      warnings?: string[];
    };
    MeshServiceItem: {
// ...
      spec: {
        identities?: {
            /** @enum {string} */
            type: "ServiceTag";
            value: string;
          }[];
        ports?: ({
            name?: string;
            /** Format: int32 */
            port: number;
            /**
             * @description Protocol identifies a protocol supported by a service.
             * @default tcp
             */
            protocol?: string;
            targetPort?: number | string;
          })[];
        selector?: {
          dataplaneRef?: {
            name?: string;
          };
          dataplaneTags?: {
            [key: string]: string;
          };
        };
      };
      /** @description Status is the current status of the Kuma MeshService resource. */
      status?: {
// ...
        vips?: {
            ip?: string;
          }[];
      };
    };
    MeshServiceCreateOrUpdateSuccessResponse: {
      /**
       * @description warnings is a list of warning messages to return to the requesting Kuma API clients.
       * Warning messages describe a problem the client making the API request should correct or be aware of.
       */
      warnings?: string[];
    };
  };
  responses: {
// ...
    MeshExternalServiceList: {
      content: {
        "application/json": {
          items?: components["schemas"]["MeshExternalServiceItem"][];
          /** @description The total number of entities */
          total?: number;
          /** @description URL to the next page */
          next?: string;
        };
      };
    };
// ...
  };
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}